# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tailor CRM API
  version: 3.0.0
  description: |
    # Production-Ready REST API for Tailor Management SaaS

    **Last Updated:** May 10, 2026
    **Status:** Fully Implemented & Tested
    **Total Endpoints:** 80 (73 + 7 system)

    ## Response Format (ALL endpoints)

    Success:   { success: true,  message: "...", data: {} }
    Error:     { success: false, message: "...", error: { code: "...", details: {} } }
    Paginated: { success: true,  message: "...", data: { items: [], pagination: { page, limit, total } } }


    ## Field Naming Convention
    - Requests:  snake_case  (customer_id, delivery_date)
    - Responses: camelCase   (customerId, deliveryDate)

    ## Order Status Flow
    pending → in_progress → ready → delivered  (cannot skip)

    ## Error Codes
    INVALID_INPUT | UNAUTHORIZED | FORBIDDEN | NOT_FOUND
    DUPLICATE_PHONE | INVALID_STATUS_TRANSITION
    CUSTOMER_NOT_FOUND | ORDER_NOT_FOUND | SHOP_NOT_FOUND
    MEASUREMENT_NOT_FOUND | STAFF_NOT_FOUND | INVOICE_NOT_FOUND
    INTERNAL_ERROR | RATE_LIMIT_EXCEEDED | SERVICE_UNAVAILABLE

    ## Rate Limiting
    All requests are rate-limited per IP address:
    - **Global**: 100 requests / 15 minutes (default)
    - **Login**: 5 requests / 15 minutes
    - **OTP/Verify**: 3 requests / 15 minutes
    - **Refresh**: 10 requests / 15 minutes
    - **Upload**: 20 requests / 15 minutes
    
    Rate limit status in response headers:
    - `RateLimit-Limit`: Maximum requests allowed
...

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        ## JWT Authentication with Token Rotation & Blacklist
        
        **Obtain tokens from:**
        - POST /auth/login (Firebase production)
        - POST /auth/login-test (Development only)
        
        **Token Structure:**
        - Access Token: 7-day expiry, contains userId, phone, shopId, sessionId, jti
        - Refresh Token: Long-lived, used for rotation
        
        **Token Rotation Behavior:**
        - Every refresh returns NEW access + refresh tokens
        - Old refresh token becomes invalid (single-use)
        - Prevents token reuse attacks
        - Session ID maintained across rotations
        
        **Security Features:**
        - Redis-backed session tracking
        - Access token blacklist on logout
        - Session activity validation
        - JTI (JWT ID) claim prevents replay attacks
        - Automatic revocation on suspicious activity
        
        **On 401 Unauthorized:**
        - Token expired? Call POST /auth/refresh-token
        - Still fails? Re-login with POST /auth/login
        - Token revoked? Session blacklisted, must re-login
        
        **Usage:**
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
security:
  - BearerAuth: []
...
paths:
...
  /staff:
    post:
      tags: [Staff]
      summary: Add staff member
      operationId: createStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                  enum: [manager, cutter, stitcher, helper, staff]
                  default: staff
                permissions:
                  type: object
                  properties:
                    viewOrders:
                      type: boolean
                    createOrders:
                      type: boolean
                    editMeasurements:
                      type: boolean
                    accessBilling:
                      type: boolean
                    accessReports:
                      type: boolean
            example:
              name: "Raju Kumar"
              phone: "+919876543211"
              role: "cutter"
              permissions:
                viewOrders: true
                createOrders: false
                editMeasurements: true
                accessBilling: false
                accessReports: false
      responses:
        '201':
          description: Staff member created
          content:
            application/json:
              example:
                success: true
                message: Staff created successfully
                data:
                  id: 1
                  shopId: 1
                  name: "Raju Kumar"
                  phone: "+919876543211"
                  role: "cutter"
                  permissions:
                    viewOrders: true
                    createOrders: false
                    editMeasurements: true
                    accessBilling: false
                    accessReports: false
                  isActive: true
                  createdAt: "2026-04-29T10:00:00Z"

    get:
      tags: [Staff]
      summary: Get all staff for shop
      operationId: getStaff
      responses:
        '200':
          description: Staff list retrieved
          content:
            application/json:
              example:
                success: true
                message: Staff retrieved successfully
                data:
                  items:
                    - id: 1
                      name: "Raju Kumar"
                      phone: "+919876543211"
                      role: "cutter"
                      isActive: true
                      permissions:
                        viewOrders: true
                        createOrders: false
                    - id: 2
                      name: "Amit Singh"
                      phone: "+919876543212"
                      role: "stitcher"
                      isActive: true
                      permissions:
                        viewOrders: true
                        createOrders: false
  /staff/{id}:
    put:
      tags: [Staff]
      summary: Update staff member
      operationId: updateStaff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                  enum: [manager, cutter, stitcher, helper, staff]
                permissions:
                  type: object
                is_active:
                  type: boolean
            example:
              role: "manager"
              is_active: true
              permissions:
                viewOrders: true
                createOrders: true
                editMeasurements: true
                accessBilling: true
                accessReports: false
      responses:
        '200':
          description: Staff updated
          content:
            application/json:
              example:
                success: true
                message: Staff updated successfully
                data:
                  id: 1
                  name: "Raju Kumar"
                  role: "manager"
                  isActive: true
                  permissions:
                    viewOrders: true
                    createOrders: true
                    editMeasurements: true
                    accessBilling: true
                    accessReports: false
                  updatedAt: "2026-04-29T14:00:00Z"

    delete:
      tags: [Staff]
      summary: Delete staff member
      operationId: deleteStaff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Staff deleted
          content:
            application/json:
              example:
                success: true
                message: Staff deleted successfully
                data: null

  # ══════════════════════════════════════════════════════
  # NOTIFICATION — 5 endpoints
  # ══════════════════════════════════════════════════════
...